requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
more-itertools==9.1.0
multidict==6.0.4
python-dotenv==1.0.0
requests==2.30.0
SQLAlchemy==1.4.48
urllib3==2.0.2
yarl==1.9.2
